add/substractQuantity
